src/edge/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --output-file requirements.txt
avro-python3==1.10.2
    # via edge
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.2.1
    # via deptry
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   deptry
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deptry==0.23.0
filelock==3.13.1
    # via torch
flatbuffers==25.2.10
    # via onnxruntime
fonttools==4.58.1
    # via matplotlib
fsspec==2024.6.1
    # via torch
humanfriendly==10.0
    # via coloredlogs
idna==3.4
    # via requests
jinja2==3.1.4
    # via torch
kafka-python==2.2.10
    # via edge
kiwisolver==1.4.8
    # via matplotlib
markupsafe==2.1.5
    # via
    #   edge
    #   jinja2
matplotlib==3.10.3
    # via ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
numpy==2.1.2
    # via
    #   contourpy
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pandas
    #   scipy
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
onnx==1.18.0
    # via
    #   edge
    #   onnxslim
onnxruntime==1.22.0
    # via edge
onnxslim==0.1.56
    # via edge
opencv-python==4.11.0.86
    # via
    #   edge
    #   ultralytics
packaging==24.1
    # via
    #   deptry
    #   matplotlib
    #   onnxruntime
    #   onnxslim
    #   requirements-parser
pandas==2.2.3
    # via ultralytics
pillow==11.0.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
protobuf==6.31.1
    # via
    #   onnx
    #   onnxruntime
psutil==7.0.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pyparsing==3.2.3
    # via matplotlib
pyreadline3==3.5.4 ; sys_platform == 'win32'
    # via humanfriendly
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via edge
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via ultralytics
requests==2.28.1
    # via ultralytics
requirements-parser==0.13.0
    # via deptry
ruff==0.11.12
scipy==1.15.3
    # via ultralytics
setuptools==70.2.0 ; python_full_version >= '3.12'
    # via torch
six==1.17.0
    # via python-dateutil
sympy==1.13.3
    # via
    #   onnxruntime
    #   onnxslim
    #   torch
torch==2.7.1 ; sys_platform == 'darwin'
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torch==2.7.1+cpu ; sys_platform != 'darwin'
    # via
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.22.1 ; (platform_machine == 'aarch64' and sys_platform == 'linux') or sys_platform == 'darwin'
    # via ultralytics
torchvision==0.22.1+cpu ; (platform_machine != 'aarch64' and sys_platform == 'linux') or (sys_platform != 'darwin' and sys_platform != 'linux')
    # via ultralytics
tqdm==4.66.5
    # via ultralytics
typing-extensions==4.12.2
    # via
    #   onnx
    #   torch
tzdata==2025.2
    # via pandas
ultralytics==8.3.149
    # via edge
ultralytics-thop==2.0.14
    # via ultralytics
urllib3==1.26.13
    # via requests
